+++ backend/controller/UserController.js
@@ -15,23 +15,29 @@ exports.get = async (req, res) => {
 };
 
 exports.update = async (req, res) => {
-  const { _id } = req.body;
+  if (!req._user.isAdmin) {
+    return res.status(401).json({ message: "You don't have permition" });
+  }
+  const _id = req.params.userId;
   const userParams = {
     username: req.body.username,
     name: req.body.name,
     password: req.body.password,
-    isAdmin: req.doby.isAdmin,
+    isAdmin: req.body.isAdmin,
   };
   await UserModel.updateOne({ _id }, { $set: userParams })
     .then(() => {
-      return res.status(200).json("User was updated!");
+      return res.status(200).json({ message: "User was updated!" });
     })
     .catch((error) => {
-      return res.status(400).json(error);
+      return res.status(400).json({ error: error });
     });
 };
 
 exports.delete = async (req, res) => {
+  if (!req._user.isAdmin) {
+    return res.status(401).json({ message: "You don't have permition" });
+  }
   const _id = req.params.userId;
   await UserModel.findByIdAndRemove({ _id })
     .then(() => {
